requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
defusedxml==0.7.1
distro==1.9.0
duckdb==1.2.1
et_xmlfile==2.0.0
fastapi==0.115.9
feedparser==6.0.11
filelock==3.18.0
fsspec==2025.3.0
fuzzywuzzy==0.18.0
h11==0.14.0
httpcore==1.0.7
httpie==3.2.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.2.0
networkx==3.4.2
nodeenv==1.9.1
numba==0.61.0
numpy==2.1.3
openai==1.69.0
openai-whisper==20240930
openpyxl==3.1.5
pandas==2.2.3
parse==1.20.2
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
pycparser==2.22
pycryptodomex==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pypdfium2==4.30.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
pyzipper==0.3.6
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
setuptools==76.1.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
svgwrite==1.4.3
sympy==1.13.1
tabula-py==2.10.0
tiktoken==0.9.0
torch==2.6.0
tqdm==4.67.1
Tree==0.2.4
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
whisper==1.1.10
